rename getDefaultStateForField.js
