+++ src/pages/ReimbursementAccount/BankAccountStep.js
@@ -30,7 +30,7 @@ import {
 } from '../../libs/actions/BankAccounts';
 import ONYXKEYS from '../../ONYXKEYS';
 import compose from '../../libs/compose';
-import getDefaultStateForField from '../../libs/getDefaultStateForField';
+import {getDefaultStateForField} from '../../libs/ReimbursementAccountUtils';
 
 const propTypes = {
     /** Bank account currently in setup */
