+++ src/pages/ReimbursementAccount/CompanyStep.js
@@ -30,7 +30,7 @@ import {
 import compose from '../../libs/compose';
 import ONYXKEYS from '../../ONYXKEYS';
 import ExpensiPicker from '../../components/ExpensiPicker';
-import getDefaultStateForField from '../../libs/getDefaultStateForField';
+import {getDefaultStateForField} from '../../libs/ReimbursementAccountUtils';
 
 const propTypes = {
     /** Bank account currently in setup */
