+++ src/pages/ReimbursementAccount/RequestorStep.js
@@ -24,7 +24,7 @@ import {isValidIdentity} from '../../libs/ValidationUtils';
 import Onfido from '../../components/Onfido';
 import compose from '../../libs/compose';
 import ONYXKEYS from '../../ONYXKEYS';
-import getDefaultStateForField from '../../libs/getDefaultStateForField';
+import {getDefaultStateForField} from '../../libs/ReimbursementAccountUtils';
 
 const propTypes = {
     /** Bank account currently in setup */
